Extract logic to CommentService in order to allow for creating custom ApiCommentController
